+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Version: 1.0.19.c
 - fix diodes not being able to transfer more than 8A
 - fix surface rocks not breaking when their supporting block is broken
 - fix crash when placing pump covers on fluid pipes in certain cases
+- fix voiding mode not working in quantum tanks